# Skip previously checked combinations in ParentGraph